+++ erasmus/json.py
@@ -2,6 +2,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Any, Final
 
+import orjson
+
 if TYPE_CHECKING:
     from collections.abc import Mapping
 
@@ -30,3 +32,11 @@ def has(obj: Mapping[str, object], key: str, /) -> bool:
     result = get(obj, key, MISSING)
 
     return result is not MISSING
+
+
+def serialize(obj: object) -> str:
+    return orjson.dumps(obj).decode('utf-8')
+
+
+def deserialize(json: str) -> object:
+    return orjson.loads(json)